+++ OneMore/Commands/File/Markdown/OneMoreDigRenderer.cs
@@ -2,88 +2,11 @@
 // Copyright © 2022 Steven M Cohn.  All rights reserved.
 //************************************************************************************************
 
-#pragma warning disable CS3001 // Argument type is not CLS-compliant
-#pragma warning disable CS3002 // Return type is not CLS-compliant
-#pragma warning disable CS3009 // Base type is not CLS-compliant
-
 namespace River.OneMoreAddIn.Commands
 {
-	using Markdig;
 	using Markdig.Renderers;
 	using Markdig.Renderers.Html;
 	using Markdig.Syntax;
-	using System;
-	using System.IO;
-
-
-	internal static class OneMoreDig
-	{
-		public static string ConvertMarkdownToHtml(string path, string markdown)
-		{
-			using var writer = new StringWriter();
-			var renderer = new HtmlRenderer(writer)
-			{
-				BaseUrl = new Uri(Path.GetDirectoryName(path) + "/")
-			};
-
-			var pipeline = new MarkdownPipelineBuilder()
-				.UseAdvancedExtensions()
-				.UseOneMoreExtensions()
-				.Build();
-
-			pipeline.Setup(renderer);
-
-			var doc = Markdown.Parse(markdown, pipeline);
-
-			renderer.Render(doc);
-			writer.Flush();
-
-			return writer.ToString();
-		}
-	}
-
-
-	internal static class OneMoreDigExtensions
-	{
-		public static MarkdownPipelineBuilder UseOneMoreExtensions(
-			this MarkdownPipelineBuilder pipeline)
-		{
-			pipeline.Extensions.Add(new OneMoreDigExtension());
-			return pipeline;
-		}
-	}
-
-
-	internal class OneMoreDigExtension : IMarkdownExtension
-	{
-		public OneMoreDigExtension()
-		{
-		}
-
-
-		public void Setup(MarkdownPipelineBuilder pipeline) { }
-
-
-		public void Setup(MarkdownPipeline pipeline, IMarkdownRenderer renderer)
-		{
-			if (renderer == null)
-			{
-				throw new ArgumentNullException(nameof(renderer));
-			}
-
-			if (renderer is TextRendererBase<HtmlRenderer> htmlRenderer)
-			{
-				var headingRenderer = htmlRenderer.ObjectRenderers.FindExact<HeadingRenderer>();
-				if (headingRenderer != null)
-				{
-					htmlRenderer.ObjectRenderers.Remove(headingRenderer);
-
-					htmlRenderer.ObjectRenderers.AddIfNotAlready(
-						new OneMoreDigRenderer(headingRenderer));
-				}
-			}
-		}
-	}
 
 
 	internal class OneMoreDigRenderer : HtmlObjectRenderer<HeadingBlock>
