+++ OneMore/OneMore.csproj
@@ -138,6 +138,9 @@
   <ItemGroup>
     <Compile Include="AddInCommands.cs" />
     <Compile Include="Commands\Edit\DuplicateLineCommand.cs" />
+    <Compile Include="Commands\File\Markdown\OneMoreDig.cs" />
+    <Compile Include="Commands\File\Markdown\OneMoreDigExtension.cs" />
+    <Compile Include="Commands\File\Markdown\OneMoreDigExtensions.cs" />
     <Compile Include="Commands\File\Markdown\OneMoreDigRenderer.cs" />
     <Compile Include="Commands\Images\OpenImageWithCommand.cs" />
     <Compile Include="Commands\Tables\ApplyTableThemeCommand.cs" />
